[LC-126] Move locations of hashing and signature.
