+++ cli_tools/icx_test/icx_wallet.py
@@ -2,14 +2,12 @@
 import base64
 import hashlib
 import logging
-import binascii
 import random
-import sys
 
 from secp256k1 import PrivateKey, PublicKey
 from loopchain import utils, configure as conf
 from loopchain.blockchain import Hash32
-from loopchain.blockchain.hashing import build_hash_generator
+from loopchain.crypto.hashing import build_hash_generator
 
 ICX_FACTOR = 10 ** 18
 ICX_FEE = 0.01
