+++ loopchain/blockchain/transactions/transaction_builder.py
@@ -3,7 +3,7 @@ import hashlib
 from abc import abstractmethod, ABC
 from typing import TYPE_CHECKING
 from .. import Signature, ExternalAddress, Hash32
-from ..hashing import build_hash_generator
+from loopchain.crypto.hashing import build_hash_generator
 
 if TYPE_CHECKING:
     from secp256k1 import PrivateKey
