+++ loopchain/crypto/hashing/hash_origin_generator.py
