+++ loopchain/peer/peer_outer_service.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 """gRPC service for Peer Outer Service"""
 import asyncio
+import copy
 import datetime
 from functools import partial
